Fix to camera sync
We were not able to sync camera trig and frame retrieval, which was
causing the TM matrix to be incorrect. I've added in some loops to
ensure each frame is triggered and grabbed, and it has been verified to
work on the chameleon femtosecond laser. It still needs further
verifications.

This also has Damien's latest additions to create a counter input for
use with the DAQ and APD.